# anime-api-client/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Anime service OpenAPI
  version: v1
servers:
  - url: http://localhost:8090
    description: Generated server url
paths:
  /v1/producers:
    get:
      tags:
        - producer-controller
      operationId: findAllProducers
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProducerGetResponse'
      security:
        - basicAuth: []
    put:
      tags:
        - producer-controller
      operationId: updateProducer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerPutRequest'
        required: true
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
    post:
      tags:
        - producer-controller
      operationId: createProducer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProducerPostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProducerGetResponse'
      security:
        - basicAuth: []
  /v1/animes:
    get:
      tags:
        - anime-controller
      summary: Get all Animes
      description: Get all animes available in system
      operationId: findAllAnimes
      parameters:
        - name: name
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of all animes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimeGetResponse'
      security:
        - basicAuth: []
    put:
      tags:
        - anime-controller
      operationId: updateAnime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimePutRequest'
        required: true
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
    post:
      tags:
        - anime-controller
      operationId: createAnime
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimePostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AnimeGetResponse'
      security:
        - basicAuth: []
  /v1/producers/{id}:
    get:
      tags:
        - producer-controller
      operationId: findProducerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProducerGetResponse'
      security:
        - basicAuth: []
    delete:
      tags:
        - producer-controller
      operationId: deleteProducerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
  /v1/animes/{id}:
    get:
      tags:
        - anime-controller
      operationId: findAnimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AnimeGetResponse'
      security:
        - basicAuth: []
    delete:
      tags:
        - anime-controller
      operationId: deleteAnimeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
      security:
        - basicAuth: []
  /v1/animes/paginated:
    get:
      tags:
        - anime-controller
      summary: Get all animes paged
      description: Get animes by page, size and/or sort
      operationId: findAllAnimesPaged
      x-spring-paginated: true
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
        - name: sort
          in: query
          description: 'Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.'
          required: false
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: Get animes paged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageAnime'
      security:
        - basicAuth: []
components:
  schemas:
    ProducerPutRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          minLength: 1
      required:
        - id
        - name
    AnimePutRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
          x-field-extra-annotation: "@jakarta.validation.constraints.NotNull(message=\"The field 'id' cannot be null\")"
        anime:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'anime' is required\")"
      required:
        - anime
        - id
    ProducerPostRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required:
        - name
    ProducerGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        name:
          type: string
    AnimePostRequest:
      type: object
      properties:
        anime:
          type: string
          minLength: 1
          x-field-extra-annotation: "@jakarta.validation.constraints.NotBlank(message=\"The field 'anime' is required\")"
      required:
        - anime
    AnimeGetResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        anime:
          type: string
    Anime:
      type: object
      properties:
        id:
          type: integer
          format: int64
        anime:
          type: string
    PageAnime:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Anime'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic